with transformCosineBlock

+++ include/discreteCosineTransformation.h
@@ -24,6 +24,6 @@
 #pragma once
 
 
-void FastDct8_transform(double vector[static 8]);
+void FastDct8_transform(float vector[static 8]);
 
-void FastDct8_inverseTransform(double vector[static 8]);
+void FastDct8_inverseTransform(float vector[static 8]);
